feat: write typescript on main; restart when changed
1. use main.development.ts instead of main.development.js
2. ignore main.development.js and main.development.js.map
