+++ package.json
@@ -152,8 +152,10 @@
     "sinon": "^4.0.0",
     "spectron": "^3.4.1",
     "style-loader": "^0.19.0",
+    "tree-kill": "^1.2.0",
     "ts-loader": "^3.1.0",
     "ts-node": "^4.1.0",
+    "tsc-watch": "^1.0.23",
     "tslint": "^5.4.3",
     "typescript": "^2.4.1",
     "url-loader": "^0.6.1",
