+++ server.js
@@ -8,6 +8,10 @@ const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 const webpackHotMiddleware = require('webpack-hot-middleware');
 const { spawn } = require('child_process');
+const tscWatch = require('tsc-watch/client');
+const treeKill = require('tree-kill');
+
+let electronProcess;
 
 const config = require('./webpack.config.development');
 
@@ -28,15 +32,33 @@ app.use(wdm);
 
 app.use(webpackHotMiddleware(compiler));
 
+function createHotElectronProcess(cb) {
+  const electronProcess = spawn('npm', ['run', 'start-hot'], { /* shell: true, */ env: process.env, stdio: 'inherit' });
+  electronProcess.on('error', spawnError => console.error(spawnError));
+  return electronProcess;
+}
+
 const server = app.listen(PORT, 'localhost', serverError => {
   if (serverError) {
     return console.error(serverError);
   }
 
   if (argv['start-hot']) {
-    spawn('npm', ['run', 'start-hot'], { shell: true, env: process.env, stdio: 'inherit' })
-      .on('close', code => process.exit(code))
-      .on('error', spawnError => console.error(spawnError));
+    tscWatch.on('first_success', () => {
+      console.log('typescript watcher ready, start electron...');
+      electronProcess = createHotElectronProcess();
+    });
+
+    tscWatch.on('subsequent_success', () => {
+      console.log('restarting electron...');
+      treeKill(electronProcess.pid, 'SIGKILL', createHotElectronProcess);
+    });
+
+    tscWatch.on('compile_errors', () => {
+      console.log('compile main error!!!!!!!!!');
+    });
+
+    tscWatch.start('-p', 'app');
   }
 
   console.log(`Listening at http://localhost:${PORT}`);
